SCAD and stl files used to produce Cogs211 Bots
These files are divided up by color and material.  Names (v*) are a bit legacy, what is important is, the files in "Orange" are printed with orange PLA or ABS (or something stronger), "Grey" is a grey color, and NinjaFlex is whatever color of flexible, NinjaFlex filament is desired for "bumper guards", meaning the springy part of the bumper that sits on top of the flexible potentiometer.